pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "apscheduler>=3.11.0",
    "google-api-python-client>=2.176.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "kaleido==0.2.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "psycopg2-binary>=2.9.10",
    "python-docx>=1.2.0",
    "python-dotenv>=1.0.1",
    "reportlab>=4.3.1",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sendgrid>=6.11.0",
    "streamlit>=1.43.1",
    "twilio>=9.5.0",
]
